[DEV] User and authentication
Created a data class that describe a user.
Created a enum that describe user roles.
Created a repository to query the user table in database.
Created a service that sign up and login a user.
Created a class that encodes a password using BCryptPasswordEncoder.
Created a configuration for web security that configure the HttpSecurity and AuthenticationManagerBuilder.