[DEV] Download URL and UUID removal
Remove UUID since we want screenshot and video of the same clip have the same name - UUID is done on frontend, relevant changes and refactoring are made accordingly.
Add getDownloadURL in storage controller to retrieve the download link given a file path.
Add isValidFilePath to verify if a file path is valid.