[DEV] User refactor
Change User to a class without implementing UserDetails, we will do that in the security module, delete any properties related to UserDetail.
One user can have multiple roles, we make a user's role to be a hash set and make a join table.
Add a class that store the roles defined (EUserRole) to the database, add relevant repository.
Add roles to RegistrationRequest as it will be defined upon registration.