[DEV] Authentication token
Token (JSON Web Token / JWT) is a "certificate" that will return to the user upon successful login and will be expired after a fixed period of time. When the user access resources, we will check the signature of the token, use its information to get user info and authenticate and authorize the user to the resource using the user's authorities.

JwtUtils is a class that helps to manipulate the token.
AuthEntryPointJwt implements the AuthenticationEntryPoint interface and override the commence function that throws an AuthenticationException when there is an unauthenticated access.
AuthTokenFilter extends OncePerRequestFilter which makes a single execution for each request to our API.